Moved FlowControl class and globals to its own file
